move tests around, remove plain invocation tests and their companions
