+++ Makefile
@@ -264,13 +264,13 @@ lint:              		  ## Run code linter to check code style
 	($(VENV_RUN); python -m pflake8 --show-source)
 
 lint-modified:     		  ## Run code linter on modified files
-	($(VENV_RUN); python -m pflake8 --show-source `git diff --name-only HEAD | grep '\.py$$' | xargs` )
+	($(VENV_RUN); python -m pflake8 --show-source `git diff --diff-filter=d --name-only HEAD | grep '\.py$$' | xargs` )
 
 format:            		  ## Run black and isort code formatter
 	($(VENV_RUN); python -m isort localstack tests; python -m black localstack tests )
 
 format-modified:   		  ## Run black and isort code formatter on modified files
-	($(VENV_RUN); python -m isort `git diff --name-only HEAD | grep '\.py$$' | xargs`; python -m black `git diff --name-only HEAD | grep '\.py$$' | xargs` )
+	($(VENV_RUN); python -m isort `git diff --diff-filter=d --name-only HEAD | grep '\.py$$' | xargs`; python -m black `git diff --diff-filter=d --name-only HEAD | grep '\.py$$' | xargs` )
 
 init-precommit:    		  ## install te pre-commit hook into your local git repository
 	($(VENV_RUN); pre-commit install)
