--- tests/integration/awslambda/functions/dotnetcore2/aws-lambda-tools-defaults.json
@@ -1,14 +1,0 @@
-{
-    "Information": [
-        "This file provides default values for the deployment wizard inside Visual Studio and the AWS Lambda commands added to the .NET Core CLI.",
-        "All the command line options for the Lambda command can be specified in this file."
-    ],
-    "profile": "default",
-    "region": "us-east-1",
-    "configuration": "Release",
-    "framework": "netcoreapp2.0",
-    "function-runtime": "dotnetcore2.0",
-    "function-memory-size": 256,
-    "function-timeout": 30,
-    "function-handler": "DotNetCore2::DotNetCore2.Lambda.Function::SimpleFunctionHandler"
-}
