+++ tests/integration/awslambda/test_lambda_integration_sqs.py
@@ -3,16 +3,23 @@ import os
 import time
 
 import pytest
+from botocore.exceptions import ClientError
 
+from localstack.services.awslambda.lambda_api import (
+    BATCH_SIZE_RANGES,
+    INVALID_PARAMETER_VALUE_EXCEPTION,
+)
 from localstack.services.awslambda.lambda_utils import (
     LAMBDA_RUNTIME_PYTHON37,
     LAMBDA_RUNTIME_PYTHON38,
+    LAMBDA_RUNTIME_PYTHON39,
 )
 from localstack.testing.aws.util import is_aws_cloud
 from localstack.utils.strings import short_uid
 from localstack.utils.sync import retry
+from localstack.utils.testutil import check_expected_lambda_log_events_length, get_lambda_log_events
 from tests.integration.awslambda.functions import lambda_integration
-from tests.integration.awslambda.test_lambda import TEST_LAMBDA_PYTHON
+from tests.integration.awslambda.test_lambda import TEST_LAMBDA_PYTHON, TEST_LAMBDA_PYTHON_ECHO
 
 THIS_FOLDER = os.path.dirname(os.path.realpath(__file__))
 LAMBDA_SQS_INTEGRATION_FILE = os.path.join(THIS_FOLDER, "functions", "lambda_sqs_integration.py")
@@ -773,3 +780,286 @@ def test_report_batch_item_failures_empty_json_batch_succeeds(
         QueueUrl=event_dlq_url, WaitTimeSeconds=retry_timeout + 1
     )
     assert "Messages" not in dlq_response
+
+
+class TestSQSEventSourceMapping:
+    # FIXME refactor and move to test_lambda_sqs_integration
+
+    @pytest.mark.skip_snapshot_verify
+    def test_event_source_mapping_default_batch_size(
+        self,
+        create_lambda_function,
+        lambda_client,
+        sqs_create_queue,
+        sqs_queue_arn,
+        lambda_su_role,
+        snapshot,
+    ):
+        snapshot.add_transformer(snapshot.transform.lambda_api())
+        function_name = f"lambda_func-{short_uid()}"
+        queue_name_1 = f"queue-{short_uid()}-1"
+        queue_name_2 = f"queue-{short_uid()}-2"
+        queue_url_1 = sqs_create_queue(QueueName=queue_name_1)
+        queue_arn_1 = sqs_queue_arn(queue_url_1)
+
+        try:
+            create_lambda_function(
+                func_name=function_name,
+                handler_file=TEST_LAMBDA_PYTHON_ECHO,
+                runtime=LAMBDA_RUNTIME_PYTHON39,
+                role=lambda_su_role,
+            )
+
+            rs = lambda_client.create_event_source_mapping(
+                EventSourceArn=queue_arn_1, FunctionName=function_name
+            )
+            snapshot.match("create-event-source-mapping", rs)
+
+            uuid = rs["UUID"]
+            assert BATCH_SIZE_RANGES["sqs"][0] == rs["BatchSize"]
+            _await_event_source_mapping_enabled(lambda_client, uuid)
+
+            with pytest.raises(ClientError) as e:
+                # Update batch size with invalid value
+                rs = lambda_client.update_event_source_mapping(
+                    UUID=uuid,
+                    FunctionName=function_name,
+                    BatchSize=BATCH_SIZE_RANGES["sqs"][1] + 1,
+                )
+            snapshot.match("invalid-update-event-source-mapping", e.value.response)
+            e.match(INVALID_PARAMETER_VALUE_EXCEPTION)
+
+            queue_url_2 = sqs_create_queue(QueueName=queue_name_2)
+            queue_arn_2 = sqs_queue_arn(queue_url_2)
+
+            with pytest.raises(ClientError) as e:
+                # Create event source mapping with invalid batch size value
+                rs = lambda_client.create_event_source_mapping(
+                    EventSourceArn=queue_arn_2,
+                    FunctionName=function_name,
+                    BatchSize=BATCH_SIZE_RANGES["sqs"][1] + 1,
+                )
+            snapshot.match("invalid-create-event-source-mapping", e.value.response)
+            e.match(INVALID_PARAMETER_VALUE_EXCEPTION)
+        finally:
+            lambda_client.delete_event_source_mapping(UUID=uuid)
+
+    @pytest.mark.aws_validated
+    def test_sqs_event_source_mapping(
+        self,
+        create_lambda_function,
+        lambda_client,
+        sqs_client,
+        sqs_create_queue,
+        sqs_queue_arn,
+        logs_client,
+        lambda_su_role,
+    ):
+        function_name = f"lambda_func-{short_uid()}"
+        queue_name_1 = f"queue-{short_uid()}-1"
+        mapping_uuid = None
+
+        try:
+            create_lambda_function(
+                func_name=function_name,
+                handler_file=TEST_LAMBDA_PYTHON_ECHO,
+                runtime=LAMBDA_RUNTIME_PYTHON37,
+                role=lambda_su_role,
+            )
+            queue_url_1 = sqs_create_queue(QueueName=queue_name_1)
+            queue_arn_1 = sqs_queue_arn(queue_url_1)
+            mapping_uuid = lambda_client.create_event_source_mapping(
+                EventSourceArn=queue_arn_1,
+                FunctionName=function_name,
+                MaximumBatchingWindowInSeconds=1,
+            )["UUID"]
+            _await_event_source_mapping_enabled(lambda_client, mapping_uuid)
+
+            sqs_client.send_message(QueueUrl=queue_url_1, MessageBody=json.dumps({"foo": "bar"}))
+
+            retry(
+                check_expected_lambda_log_events_length,
+                retries=10,
+                sleep=1,
+                function_name=function_name,
+                expected_length=1,
+                logs_client=logs_client,
+            )
+
+            rs = sqs_client.receive_message(QueueUrl=queue_url_1)
+            assert rs.get("Messages") is None
+        finally:
+            if mapping_uuid:
+                lambda_client.delete_event_source_mapping(UUID=mapping_uuid)
+
+    @pytest.mark.aws_validated
+    @pytest.mark.parametrize(
+        "filter, item_matching, item_not_matching",
+        [
+            # test single filter
+            (
+                {"body": {"testItem": ["test24"]}},
+                {"testItem": "test24"},
+                {"testItem": "tesWER"},
+            ),
+            # test OR filter
+            (
+                {"body": {"testItem": ["test24", "test45"]}},
+                {"testItem": "test45"},
+                {"testItem": "WERTD"},
+            ),
+            # test AND filter
+            (
+                {"body": {"testItem": ["test24", "test45"], "test2": ["go"]}},
+                {"testItem": "test45", "test2": "go"},
+                {"testItem": "test67", "test2": "go"},
+            ),
+            # exists
+            (
+                {"body": {"test2": [{"exists": True}]}},
+                {"test2": "7411"},
+                {"test5": "74545"},
+            ),
+            # numeric (bigger)
+            (
+                {"body": {"test2": [{"numeric": [">", 100]}]}},
+                {"test2": 105},
+                "this is a test string",  # normal string should be dropped as well aka not fitting to filter
+            ),
+            # numeric (smaller)
+            (
+                {"body": {"test2": [{"numeric": ["<", 100]}]}},
+                {"test2": 93},
+                {"test2": 105},
+            ),
+            # numeric (range)
+            (
+                {"body": {"test2": [{"numeric": [">=", 100, "<", 200]}]}},
+                {"test2": 105},
+                {"test2": 200},
+            ),
+            # prefix
+            (
+                {"body": {"test2": [{"prefix": "us-1"}]}},
+                {"test2": "us-1-48454"},
+                {"test2": "eu-wert"},
+            ),
+        ],
+    )
+    def test_sqs_event_filter(
+        self,
+        create_lambda_function,
+        lambda_client,
+        sqs_client,
+        sqs_create_queue,
+        sqs_queue_arn,
+        logs_client,
+        lambda_su_role,
+        filter,
+        item_matching,
+        item_not_matching,
+    ):
+        function_name = f"lambda_func-{short_uid()}"
+        queue_name_1 = f"queue-{short_uid()}-1"
+        mapping_uuid = None
+
+        try:
+            create_lambda_function(
+                func_name=function_name,
+                handler_file=TEST_LAMBDA_PYTHON_ECHO,
+                runtime=LAMBDA_RUNTIME_PYTHON37,
+                role=lambda_su_role,
+            )
+            queue_url_1 = sqs_create_queue(QueueName=queue_name_1)
+            queue_arn_1 = sqs_queue_arn(queue_url_1)
+
+            sqs_client.send_message(QueueUrl=queue_url_1, MessageBody=json.dumps(item_matching))
+            sqs_client.send_message(
+                QueueUrl=queue_url_1,
+                MessageBody=json.dumps(item_not_matching)
+                if not isinstance(item_not_matching, str)
+                else item_not_matching,
+            )
+
+            def _assert_qsize():
+                response = sqs_client.get_queue_attributes(
+                    QueueUrl=queue_url_1, AttributeNames=["ApproximateNumberOfMessages"]
+                )
+                assert int(response["Attributes"]["ApproximateNumberOfMessages"]) == 2
+
+            retry(_assert_qsize, retries=10)
+
+            mapping_uuid = lambda_client.create_event_source_mapping(
+                EventSourceArn=queue_arn_1,
+                FunctionName=function_name,
+                MaximumBatchingWindowInSeconds=1,
+                FilterCriteria={
+                    "Filters": [
+                        {"Pattern": json.dumps(filter)},
+                    ]
+                },
+            )["UUID"]
+            _await_event_source_mapping_enabled(lambda_client, mapping_uuid)
+
+            def _check_lambda_logs():
+                events = get_lambda_log_events(function_name, logs_client=logs_client)
+                # once invoked
+                assert len(events) == 1
+                records = events[0]["Records"]
+                # one record processed
+                assert len(records) == 1
+                # check for correct record presence
+                if "body" in json.dumps(filter):
+                    item_matching_str = json.dumps(item_matching)
+                    assert records[0]["body"] == item_matching_str
+
+            retry(_check_lambda_logs, retries=10)
+
+            rs = sqs_client.receive_message(QueueUrl=queue_url_1)
+            assert rs.get("Messages") is None
+
+        finally:
+            if mapping_uuid:
+                lambda_client.delete_event_source_mapping(UUID=mapping_uuid)
+
+    @pytest.mark.aws_validated
+    @pytest.mark.parametrize(
+        "invalid_filter", [None, "simple string", {"eventSource": "aws:sqs"}, {"eventSource": []}]
+    )
+    def test_sqs_invalid_event_filter(
+        self,
+        create_lambda_function,
+        sqs_create_queue,
+        sqs_queue_arn,
+        lambda_su_role,
+        lambda_client,
+        invalid_filter,
+    ):
+        function_name = f"lambda_func-{short_uid()}"
+        queue_name_1 = f"queue-{short_uid()}"
+
+        create_lambda_function(
+            func_name=function_name,
+            handler_file=TEST_LAMBDA_PYTHON_ECHO,
+            runtime=LAMBDA_RUNTIME_PYTHON37,
+            role=lambda_su_role,
+        )
+        queue_url_1 = sqs_create_queue(QueueName=queue_name_1)
+        queue_arn_1 = sqs_queue_arn(queue_url_1)
+
+        with pytest.raises(Exception) as expected:
+            lambda_client.create_event_source_mapping(
+                EventSourceArn=queue_arn_1,
+                FunctionName=function_name,
+                MaximumBatchingWindowInSeconds=1,
+                FilterCriteria={
+                    "Filters": [
+                        {
+                            "Pattern": invalid_filter
+                            if isinstance(invalid_filter, str)
+                            else json.dumps(invalid_filter)
+                        },
+                    ]
+                },
+            )
+        expected.match(INVALID_PARAMETER_VALUE_EXCEPTION)
